+++ src/collection/find.ts
@@ -1,0 +1,4 @@
+import { CommandCursor } from "../protocol/mod.ts";
+
+export class FindCursor<T> extends CommandCursor<T> {
+}
