+++ src/collection/mod.ts
@@ -1,0 +1,1 @@
+export { Collection } from "./collection.ts";
